修复bug

+++ jsonx/int.go
@@ -23,6 +23,6 @@ func (re JInt) ToObjPtr() *JObj   { obj := re.ToObj(); return &obj }
 func (re JInt) ToArr() JArr       { return JArr{} }
 func (re JInt) ToJDoc() JDoc      { return JDoc(re.String()) }
 func (re JInt) ToJVal() JValue    { return re }
-func (re JInt) ToGVal() any       { return float64(re) }
+func (re JInt) ToGVal() any       { return int64(re) }
 
 func NewJInt(val int64) JInt { return JInt(val) }
